+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/config/ConfigManager.java
@@ -6,88 +6,114 @@ import com.tristankechlo.random_mob_sizes.IPlatformHelper;
 import com.tristankechlo.random_mob_sizes.RandomMobSizes;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.FileWriter;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.List;
 
 public final class ConfigManager {
 
     private static final File CONFIG_DIR = IPlatformHelper.INSTANCE.getConfigDirectory().toFile();
-    public static final Gson GSON = new GsonBuilder().setPrettyPrinting().serializeNulls().create();
+    private static final Gson GSON = new GsonBuilder().setPrettyPrinting().serializeNulls().create();
     public static final String FILE_NAME = RandomMobSizes.MOD_ID + ".json";
+    private static final File CONFIG_FILE = new File(CONFIG_DIR, FILE_NAME);
 
-    public static void loadAndVerifyConfig() {
+    public static boolean loadAndVerifyConfig() {
         ConfigManager.createConfigFolder();
         RandomMobSizesConfig.setToDefault();
-        File configFile = new File(CONFIG_DIR, FILE_NAME);
-        if (configFile.exists()) {
-            ConfigManager.loadConfigFromFile(configFile);
-            ConfigManager.writeConfigToFile(configFile);
-            RandomMobSizes.LOGGER.info("Saved the checked/corrected config: '{}'", FILE_NAME);
+        if (CONFIG_FILE.exists()) {
+            try {
+                ConfigManager.loadConfigFromFile();
+                RandomMobSizes.LOGGER.info("Config '{}' was successfully loaded.", FILE_NAME);
+                return true;
+            } catch (Exception e) {
+                RandomMobSizes.LOGGER.error(e.getMessage());
+                RandomMobSizes.LOGGER.error("Error loading config '{}', config hasn't been loaded. Using default config.", FILE_NAME);
+                ConfigManager.backupConfig(); // save content of old config to a backup file
+                RandomMobSizesConfig.setToDefault();
+                ConfigManager.writeConfigToFile(); // write default config to file
+                return false;
+            }
         } else {
-            ConfigManager.writeConfigToFile(configFile);
+            ConfigManager.writeConfigToFile();
             RandomMobSizes.LOGGER.warn("No config '{}' was found, created a new one.", FILE_NAME);
+            return true;
         }
     }
 
-    private static void writeConfigToFile(File file) {
-        JsonElement jsonObject = RandomMobSizesConfig.serialize(new JsonObject());
+    private static void loadConfigFromFile() throws FileNotFoundException {
+        JsonParser parser = new JsonParser();
+        JsonElement jsonElement = parser.parse(new FileReader(CONFIG_FILE));
+        JsonObject json = jsonElement.getAsJsonObject();
+        RandomMobSizesConfig.deserialize(json);
+    }
+
+    private static void writeConfigToFile() {
         try {
-            JsonWriter writer = new JsonWriter(new FileWriter(file));
+            JsonElement jsonObject = RandomMobSizesConfig.serialize();
+            JsonWriter writer = new JsonWriter(new FileWriter(CONFIG_FILE));
             writer.setIndent("\t");
             GSON.toJson(jsonObject, writer);
             writer.close();
         } catch (Exception e) {
             RandomMobSizes.LOGGER.error("There was an error writing the config to file: '{}'", FILE_NAME);
-            e.printStackTrace();
+            RandomMobSizes.LOGGER.error(e.getMessage());
         }
     }
 
-    private static void loadConfigFromFile(File file) {
-        JsonObject json = null;
+    private static void backupConfig() {
+        String backupFileName = FILE_NAME.replace(".json", ".backup.txt");
+        Path backupFilePath = Paths.get(CONFIG_DIR.getAbsolutePath(), backupFileName);
         try {
-            JsonParser parser = new JsonParser();
-            JsonElement jsonElement = parser.parse(new FileReader(file));
-            json = jsonElement.getAsJsonObject();
+            List<String> lines = Files.readAllLines(CONFIG_FILE.toPath());
+            Files.write(backupFilePath, lines);
+            RandomMobSizes.LOGGER.warn("Created backup file '{}'", backupFileName);
         } catch (Exception e) {
-            RandomMobSizes.LOGGER.error("There was an error loading the config file: '{}'", FILE_NAME);
-            e.printStackTrace();
-        }
-        if (json != null) {
-            RandomMobSizesConfig.deserialize(json);
-            RandomMobSizes.LOGGER.info("Config '{}' was successfully loaded.", FILE_NAME);
-        } else {
-            RandomMobSizes.LOGGER.error("Error loading config '{}', config hasn't been loaded.", FILE_NAME);
+            RandomMobSizes.LOGGER.error("Error creating backup file '{}'", backupFileName);
+            RandomMobSizes.LOGGER.error(e.getMessage());
         }
     }
 
-    public static void resetConfig() {
-        RandomMobSizesConfig.setToDefault();
-        File configFile = new File(CONFIG_DIR, FILE_NAME);
-        ConfigManager.writeConfigToFile(configFile);
-        RandomMobSizes.LOGGER.info("Config '{}' was set to default.", FILE_NAME);
-    }
-
-    public static void saveConfig() {
-        File configFile = new File(CONFIG_DIR, FILE_NAME);
-        ConfigManager.writeConfigToFile(configFile);
-        RandomMobSizes.LOGGER.info("Config '{}' was saved.", FILE_NAME);
-    }
-
-    public static void reloadConfig() {
-        File configFile = new File(CONFIG_DIR, FILE_NAME);
-        if (configFile.exists()) {
-            ConfigManager.loadConfigFromFile(configFile);
-            ConfigManager.writeConfigToFile(configFile);
-            RandomMobSizes.LOGGER.info("Saved the checked/corrected config: " + FILE_NAME);
-        } else {
-            ConfigManager.writeConfigToFile(configFile);
-            RandomMobSizes.LOGGER.warn("No config '{}' was found, created a new one.", FILE_NAME);
+    public static boolean resetConfig() {
+        try {
+            RandomMobSizesConfig.setToDefault();
+            ConfigManager.writeConfigToFile();
+            RandomMobSizes.LOGGER.info("Config '{}' was set to default.", FILE_NAME);
+            return true;
+        } catch (Exception e) {
+            RandomMobSizes.LOGGER.error("Error resetting config '{}'", FILE_NAME);
+            RandomMobSizes.LOGGER.error(e.getMessage());
         }
+        return false;
+    }
+
+    public static boolean saveConfig() {
+        try {
+            ConfigManager.writeConfigToFile();
+            RandomMobSizes.LOGGER.info("Config '{}' was saved.", FILE_NAME);
+            return true;
+        } catch (Exception e) {
+            RandomMobSizes.LOGGER.error("Error saving config '{}'", FILE_NAME);
+            RandomMobSizes.LOGGER.error(e.getMessage());
+        }
+        return false;
+    }
+
+    public static boolean reloadConfig() {
+        try {
+            return ConfigManager.loadAndVerifyConfig();
+        } catch (Exception e) {
+            RandomMobSizes.LOGGER.error("Error reloading config '{}'", FILE_NAME);
+            RandomMobSizes.LOGGER.error(e.getMessage());
+        }
+        return false;
     }
 
     public static String getConfigPath() {
-        File configFile = new File(CONFIG_DIR, FILE_NAME);
-        return configFile.getAbsolutePath();
+        return CONFIG_FILE.getAbsolutePath();
     }
 
     private static void createConfigFolder() {