+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/config/ConfigParseException.java
@@ -1,0 +1,9 @@
+package com.tristankechlo.random_mob_sizes.config;
+
+public class ConfigParseException extends RuntimeException {
+
+    public ConfigParseException(String message) {
+        super(message);
+    }
+
+}