+++ Forge/src/main/java/com/tristankechlo/random_mob_sizes/ForgeRandomMobSizes.java
@@ -5,22 +5,19 @@ import com.tristankechlo.random_mob_sizes.commands.RandomMobSizesCommand;
 import com.tristankechlo.random_mob_sizes.config.ConfigManager;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.RegisterCommandsEvent;
-import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.event.server.ServerStartingEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 @Mod(RandomMobSizes.MOD_ID)
 public class ForgeRandomMobSizes {
 
     public ForgeRandomMobSizes() {
-        IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
-        modEventBus.addListener(this::commonSetup);
+        MinecraftForge.EVENT_BUS.addListener(this::onServerStart);
         MinecraftForge.EVENT_BUS.addListener(this::registerCommands);
         MinecraftForge.EVENT_BUS.register(this);
     }
 
-    private void commonSetup(final FMLCommonSetupEvent event) {
+    private void onServerStart(final ServerStartingEvent event) {
         ConfigManager.loadAndVerifyConfig();
     }